pyproject.toml
[tool.poetry]
name = "districkt-summer-practice-2024"
version = "0.1.0"
description = "Periodic parsing of the MTS Link API"
authors = ["Vitaly Zorin <addefan@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
python-dateutil = "^2.9.0.post0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
